renamed flags test file; updated flag tests for generateFlagOverrides and applyFlagOverrides - no longer need to mock store
